+++ tests/rest/checkjob_tester.py
@@ -25,7 +25,7 @@ class CheckjobTester:
 
         self.request_data = {
             'sid': None,
-            'original_spl': None,
+            'original_otl': None,
             'tws': 0,
             'twf': 0,
             'cache_ttl': 60,
@@ -43,8 +43,8 @@ class CheckjobTester:
         self.db.execute_query(query_str, with_commit=True)
 
     def get_jobs_from_db(self, limit=None):
-        query_str = f"""SELECT id, creating_date, status FROM splqueries 
-        WHERE original_spl='{self.original_spl}' ORDER BY creating_date DESC;"""
+        query_str = f"""SELECT id, creating_date, status FROM OTLQueries 
+        WHERE original_otl='{self.original_otl}' ORDER BY creating_date DESC;"""
         if limit:
             query_str = query_str.replace(';', f' LIMIT {limit};')
 
@@ -58,33 +58,33 @@ class CheckjobTester:
             return self.db.cur.execute(query_str, fetchall=True)
 
     def update_job_status(self, status, job_id):
-        query_str = f"""UPDATE splqueries SET status='{status}' WHERE id={job_id};"""
+        query_str = f"""UPDATE OTLQueries SET status='{status}' WHERE id={job_id};"""
         self.db.execute_query(query_str, with_commit=True, with_fetch=False)
 
     def add_job_with_cache(self):
         self.tick_dispatcher()
-        job_id, _ = self.db.add_job(search=[self.original_spl, self._current_spl], subsearches=[], tws=0,
+        job_id, _ = self.db.add_job(search=[self.original_otl, self._current_spl], subsearches=[], tws=0,
                                     twf=0, cache_ttl=60, username='tester', field_extraction=False, preview=False)
-        self.db.add_to_cache(original_spl=self.original_spl, tws=0, twf=0, cache_id=job_id, expiring_date=1)
+        self.db.add_to_cache(original_otl=self.original_otl, tws=0, twf=0, cache_id=job_id, expiring_date=1)
         return job_id
 
     def add_job_with_status(self, status):
         self.tick_dispatcher()
-        job_id, _ = self.db.add_job(search=[self.original_spl, self.original_spl], subsearches=[], tws=0, twf=0,
+        job_id, _ = self.db.add_job(search=[self.original_otl, self.original_otl], subsearches=[], tws=0, twf=0,
                                     cache_ttl=60, username='tester', field_extraction=False, preview=False)
         self.update_job_status(status, job_id)
 
     @property
-    def original_spl(self):
-        original_spl = re.sub(r"\|\s*ot\s[^|]*\|", "", self._current_spl)
-        original_spl = re.sub(r"\|\s*simple[^\"]*", "", original_spl)
-        original_spl = original_spl.replace("oteval", "eval")
-        original_spl = original_spl.strip()
-        return original_spl
+    def original_otl(self):
+        original_otl = re.sub(r"\|\s*ot\s[^|]*\|", "", self._current_spl)
+        original_otl = re.sub(r"\|\s*simple[^\"]*", "", original_otl)
+        original_otl = original_otl.replace("oteval", "eval")
+        original_otl = original_otl.strip()
+        return original_otl
 
     def send_request(self):
         data = self.request_data
-        data['original_spl'] = self._current_spl
+        data['original_otl'] = self._current_spl
         data['sid'] = str(uuid.uuid4())
         resp = requests.get(f'http://{self.config["host"]}:{self.config["port"]}/api/checkjob', params=data)
         resp.raise_for_status()
@@ -92,8 +92,8 @@ class CheckjobTester:
 
     def _cleanup(self):
         del_ticks_query = f"""DELETE FROM Ticks WHERE applicationId='test_app';"""
-        del_spl_query = f"""DELETE FROM splqueries WHERE original_spl='{self.original_spl}';"""
-        del_cache_query = f"""DELETE FROM cachesdl WHERE original_spl='{self.original_spl}';"""
+        del_spl_query = f"""DELETE FROM OTLQueries WHERE original_otl='{self.original_otl}';"""
+        del_cache_query = f"""DELETE FROM cachesdl WHERE original_otl='{self.original_otl}';"""
         for query in [del_cache_query, del_ticks_query, del_spl_query]:
             self.db.execute_query(query, with_commit=True, with_fetch=False)
 
@@ -108,7 +108,7 @@ class CheckjobTester:
     def test__new(self):
         try:
             self.tick_dispatcher()
-            self.db.add_job(search=[self.original_spl, self.original_spl], subsearches=[], tws=0, twf=0,
+            self.db.add_job(search=[self.original_otl, self.original_otl], subsearches=[], tws=0, twf=0,
                             cache_ttl=60, username='tester', field_extraction=False, preview=False)
             resp = self.send_request()
         finally: